130/146

+++ src/8.execution/utils.c
@@ -25,12 +25,49 @@ int	is_builtin(const char *cmd)
 
 void	execute_builtin_child(t_minishell *mshell, t_execution *exec)
 {
-		execute_builtin(mshell, exec->cmd_args);
+	if (is_builtin(exec->cmd_args[0]))
+	{
+		mshell->exitcode = execute_builtin(mshell, exec->cmd_args);
 		delete_minishell(mshell);
 		exit(mshell->exitcode);
-
+	}
 }
 
+int	redirect_and_restore_builtin(t_execution *exec,
+	t_minishell *mshell, char **cmd_args)
+	{
+	int origin_stdin = -1;
+	int origin_stdout = -1;
+
+	if (save_and_redirect_stdin(exec->redir_fd[FD_IN], &origin_stdin) == FAIL)
+	return (FAIL);
+
+	if (save_and_redirect_stdout(exec->redir_fd[FD_OUT], &origin_stdout) == FAIL)
+	{
+	restore_stdin(origin_stdin);
+	return (FAIL);
+	}
+
+	// Run builtin and store its exit code
+	mshell->exitcode = execute_builtin(mshell, cmd_args);
+	restore_stdout(origin_stdout);
+	restore_stdin(origin_stdin);
+
+	return (SUCCESS);
+}
+int execute_builtin_parent(t_minishell *mshell, t_execution *exec)
+{
+	if (is_builtin(exec->cmd_args[0]) && !exec->has_pipe)
+	{
+		if (redirect_and_restore_builtin(exec, mshell, exec->cmd_args) == FAIL)
+		{
+			mshell->exitcode = 1;
+			return FAIL;
+		}
+		return SUCCESS;
+	}
+	return FAIL;
+}
 
 void handle_parent(t_minishell *mshell, t_execution *exec, int *pipefd, pid_t pid)
 {
@@ -50,42 +87,42 @@ void handle_parent(t_minishell *mshell, t_execution *exec, int *pipefd, pid_t pi
 }
 
 
-int execute_builtin_parent(t_minishell *mshell, t_execution *exec)
-{
-	int original_stdin = dup(STDIN_FILENO);
-	int original_stdout = dup(STDOUT_FILENO);
-	int status;
+// int execute_builtin_parent(t_minishell *mshell, t_execution *exec)
+// {
+// 	int original_stdin = dup(STDIN_FILENO);
+// 	int original_stdout = dup(STDOUT_FILENO);
+// 	int status;
 
-	if (original_stdin == -1 || original_stdout == -1)
-	{
-		perror("dup");
-		return (1);
-	}
+// 	if (original_stdin == -1 || original_stdout == -1)
+// 	{
+// 		perror("dup");
+// 		return (1);
+// 	}
 
-	// Apply redirections
-	if (exec->redir_fd[FD_IN] != STDIN_FILENO)
-	{
-		if (dup2(exec->redir_fd[FD_IN], STDIN_FILENO) == -1)
-			perror("dup2 in");
-		close(exec->redir_fd[FD_IN]);
-	}
-	if (exec->redir_fd[FD_OUT] != STDOUT_FILENO)
-	{
-		if (dup2(exec->redir_fd[FD_OUT], STDOUT_FILENO) == -1)
-			perror("dup2 out");
-		close(exec->redir_fd[FD_OUT]);
-	}
+// 	// Apply redirections
+// 	if (exec->redir_fd[FD_IN] != STDIN_FILENO)
+// 	{
+// 		if (dup2(exec->redir_fd[FD_IN], STDIN_FILENO) == -1)
+// 			perror("dup2 in");
+// 		close(exec->redir_fd[FD_IN]);
+// 	}
+// 	if (exec->redir_fd[FD_OUT] != STDOUT_FILENO)
+// 	{
+// 		if (dup2(exec->redir_fd[FD_OUT], STDOUT_FILENO) == -1)
+// 			perror("dup2 out");
+// 		close(exec->redir_fd[FD_OUT]);
+// 	}
 
-	// Execute the builtin command in parent
-	status = execute_builtin(mshell, exec->cmd_args); // You must implement run_builtin()
+// 	// Execute the builtin command in parent
+// 	status = execute_builtin(mshell, exec->cmd_args); // You must implement run_builtin()
 
-	// Restore original stdio
-	dup2(original_stdin, STDIN_FILENO);
-	dup2(original_stdout, STDOUT_FILENO);
-	close(original_stdin);
-	close(original_stdout);
+// 	// Restore original stdio
+// 	dup2(original_stdin, STDIN_FILENO);
+// 	dup2(original_stdout, STDOUT_FILENO);
+// 	close(original_stdin);
+// 	close(original_stdout);
 
-	// Set shell's exit code
-	mshell->exitcode = status;
-	return (status);
-}
+// 	// Set shell's exit code
+// 	mshell->exitcode = status;
+// 	return (status);
+// }
